Showing a bad fix for chunk problem
This appends the portion of the chunk that isn't matched to the next chunk. If everything that's passed in is json objects, this means that if one gets cut up between two chunks, the first part of the broken object is added to the beginning of the next chunk when it's passed in.